Rename LockFile to LockFileV7 and move it to lockfile-v7.hh
In preparation for the version 8 sparse lock file format, rename
LockFile to LockFileV7 and move it, together with the Node and
LockedNode data types, to a new header lockfile-v7.hh (implementation
in lockfile-v7.cc). lockfile.hh continues to contain the common types
like InputAttrPath and NonEmptyInputAttrPath.

No behaviour change.
