+++ src/libflake/include/nix/flake/meson.build
@@ -5,6 +5,7 @@ include_dirs = [ include_directories('../..') ]
 headers = files(
   'flake.hh',
   'flakeref.hh',
+  'lockfile-v7.hh',
   'lockfile.hh',
   'provenance.hh',
   'settings.hh',
